requirements.txt
anyio==3.7.0
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
docker==6.1.3
fastapi==0.98.0
h11==0.14.0
httptools==0.5.0
idna==3.4
loguru==0.7.0
packaging==23.1
pip==23.1.2
pydantic==1.10.9
python-dotenv==1.0.0
pywin32==306
PyYAML==6.0
requests==2.31.0
setuptools==65.5.1
sniffio==1.3.0
starlette==0.27.0
typing-extensions==4.6.3
urllib3==2.0.3
uvicorn==0.22.0
watchfiles==0.19.0
websocket-client==1.6.1
websockets==11.0.3
wheel==0.38.4
win32-setctime==1.1.0